Add files via upload
From the outlook of the credit market, this
project aims at a comparative analysis of the nature and aspect
of the credit market. The project identifies the credit rationing and
assumes the lack of acceptable instruments that could be used by
banks to separately select borrowers. The project points out the
scope of individuals in the credit market, this project is mathematically described. Even if the mathematical confirmation of the
possible performance of the credit model is included in this project.